% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{aes_to_mtx}
\alias{aes_to_mtx}
\title{Helper function}
\usage{
aes_to_mtx(mapping)
}
\arguments{
\item{mapping}{TODO: aesthetics mapping?}
}
\value{
a dataframe of aes:variables
}
\description{
Helper function
}
