% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf.R
\name{combine_pmfs}
\alias{combine_pmfs}
\title{tries to merge a list of pmfs}
\usage{
combine_pmfs(res)
}
\arguments{
\item{res}{a list of expressions like A, B, C}
}
\value{
a list of
}
\description{
tries to merge a list of pmfs
}
