% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{complete_conditionals}
\alias{complete_conditionals}
\title{Helper function: add in the implied P(1|A), etc}
\usage{
complete_conditionals(prob_mtx)
}
\arguments{
\item{prob_mtx}{probability matrix as specified}
}
\value{
the matrix with probability aesthetics, including
}
\description{
Helper function: add in the implied P(1|A), etc
}
