% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{elements_checker}
\alias{elements_checker}
\title{Helper: get all the elements in the height and combine into one vector, compare two sets
all elements in coord_aes should be included in the prob_aes}
\usage{
elements_checker(prob_aes, coord_aes)
}
\arguments{
\item{prob_aes, }{coord_aes}
}
\value{
error message or nothing
}
\description{
Helper: get all the elements in the height and combine into one vector, compare two sets
all elements in coord_aes should be included in the prob_aes
}
