% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{flatten_aes}
\alias{flatten_aes}
\title{helper function}
\usage{
flatten_aes(mapping)
}
\arguments{
\item{mapping:}{list $x = [], $width = []}
}
\value{
flattened mapping: $x1 = , $width1 = , $width2 =
}
\description{
helper function
}
