% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-icon.R
\name{geom_icon}
\alias{geom_icon}
\title{From geom_point}
\usage{
geom_icon(mapping = NULL, data = NULL, stat = "icon", ...,
  na.rm = FALSE, position = "identity", show.legend = NA,
  inherit.aes = TRUE)
}
\description{
From geom_point
}
