% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_all_rv}
\alias{get_all_rv}
\title{Helper: returns the names of all random variables contained in a parsed
mapping.}
\usage{
get_all_rv(mapping)
}
\arguments{
\item{mapping}{mapping parsed by `parse_aes`}
}
\value{
vector of strings containing random variable names
}
\description{
Helper: returns the names of all random variables contained in a parsed
mapping.
}
