% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{mtx_check}
\alias{mtx_check}
\title{Helper func: Goes thru the matrix to check if it's a legit factorization of a pmf}
\usage{
mtx_check(m)
}
\arguments{
\item{mtx}{rows: pmf terms, cols: marginals, conditionals, aesthetics names}
}
\description{
Helper func: Goes thru the matrix to check if it's a legit factorization of a pmf
}
