% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_aes}
\alias{parse_aes}
\title{parses and checks aesthetics mapping}
\usage{
parse_aes(mapping)
}
\arguments{
\item{mapping}{aesthetics mapping, a list of expressions (used to be quosures)}
}
\description{
parses and checks aesthetics mapping
}
