% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf.R
\name{parse_pmf}
\alias{parse_pmf}
\title{parses a single pmf}
\usage{
parse_pmf(p)
}
\arguments{
\item{p}{a single expression like P(A, B | C, D)}
}
\value{
a list of marginals and conditionals
}
\description{
parses a single pmf
}
