% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf.R
\name{parse_pmfs}
\alias{parse_pmfs}
\title{each pmf is an expression like `P(A, B| C)`}
\usage{
parse_pmfs(one_d = NULL, two_d = NULL)
}
\arguments{
\item{length}{a list of pmfs assigned to the length aesthetics}

\item{area}{a list of pmfs assigned to the area aesthetics}
}
\value{
marginals and conditionals
}
\description{
each pmf is an expression like `P(A, B| C)`
}
