% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{print.prob_chain}
\alias{print.prob_chain}
\title{prints pretty version of the aesthetics mapping}
\usage{
\method{print}{prob_chain}(aes_mtx)
}
\description{
prints pretty version of the aesthetics mapping
}
