% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_fakecount.R
\name{stat_fakecount}
\alias{stat_fakecount}
\title{A geom that doesn't work}
\usage{
stat_fakecount(mapping = NULL, data = NULL, geom = "bar",
  position = "stack", ..., width = NULL, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)
}
\value{
a geom
}
\description{
A geom that doesn't work
}
