% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatConditionalProb.R
\docType{data}
\name{StatConditionalProb}
\alias{StatConditionalProb}
\title{New stat?}
\format{An object of class \code{StatCount} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
StatConditionalProb
}
\value{
a stat
}
\description{
New stat?
}
\keyword{datasets}
