% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-icon.R
\docType{data}
\name{StatIcon}
\alias{StatIcon}
\title{calculates the x, y coordiantes for the icon array}
\format{An object of class \code{StatIcon} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
StatIcon
}
\description{
calculates the x, y coordiantes for the icon array
}
\keyword{datasets}
