% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{flatten_aes}
\alias{flatten_aes}
\title{helper function: number those aesthetics since}
\usage{
flatten_aes(mapping)
}
\arguments{
\item{mapping:}{list $x = c(A), $width = c(P(B|A), P(A))}
}
\value{
flattened mapping: $width1 = , $width2 =
}
\description{
helper function: number those aesthetics since
}
