% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-icon.R
\docType{data}
\name{StatIcon}
\alias{StatIcon}
\title{calculates the x, y aesthetics so that GeomPoint will be happy}
\format{An object of class \code{StatIcon} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 6.}
\usage{
StatIcon
}
\description{
calculates the x, y aesthetics so that GeomPoint will be happy
}
\keyword{datasets}
